pyoverload II; unittests; minor fixes
Upgraded pyoverload into Version II (accelerated, lower-cased types, simplified); Added unit tests using package unittest; improvement in batorch and micomputing

+++ packing/pack.py
@@ -1,18 +1,25 @@
 
 from pycamia import info_manager
+
 __info__ = info_manager(
     project = "PyCAMIA",
     fileinfo = "File to pack packages in the project.",
     help = "Use `python pack.py packagename` to pack and upload packages. ",
-    requires = ['twine', 'wheel']
+    requires = ['twine', 'wheel', 'pexpect']
 ).check()
 
-import sys, os
+import sys, os, time, pexpect
+from subprocess import PIPE, Popen
 
-from pycamia import touch, Error, enclosed_object, to_list
-from import_all import import_all
+with __info__:
+    from pycamia.manager import update_format
+    from pycamia import touch, Error, enclosed_object, to_list, is_valid_command
+    from import_all import import_all
+    from check_import import check_import
 
-key_chain = {"usr": "None", "psw": "None"}
+usr, psw, *_ = open(os.path.join(os.pardir, os.pardir, "pypi.keychain")).read().split()
+key_chain = {"usr": usr, "psw": psw}
+# key_chain = {"usr": "None", "psw": "None"} # One can use plain text here.
 home_path = os.pardir
 packing_package_path = os.path.join(os.curdir, "packing_package")
 pack_type_command = dict(
@@ -23,6 +30,22 @@ pack_type_command = dict(
 if os.path.exists("github_main.url"):
     github_mainpage = open("github_main.url").read().strip()
 else: github_mainpage = ""
+update_code_list = []
+
+def find_last_modification(cpath):
+    ctime = os.path.getmtime(cpath)
+    if not os.path.isdir(cpath): return ctime
+    for f in os.listdir(cpath):
+        if f.startswith('__'): continue
+        if f.startswith('.'): continue
+        npath = os.path.join(cpath, f)
+        if not os.path.isdir(npath): ntime = os.path.getmtime(npath)
+        else: ntime = find_last_modification(npath)
+        if ntime > ctime:
+            ctime = ntime
+            
+    return ctime
+
 def do_package(name, version=None, pack_type="zipwheel", upload=True):
     """
     Setup, pack and upload the package. 
@@ -32,25 +55,36 @@ def do_package(name, version=None, pack_type="zipwheel", upload=True):
         pack_type[str]: pack to zip or wheel, both are commonly used
         upload[bool]: whether to upload
     """
-    # Move package to the temporary working directory
-    import_all([name])
-    os.system(f"cp -r {os.path.join(home_path, name)} {os.path.join(packing_package_path, name)}")
-    os.system(f"cp -r {os.path.join(home_path, name, 'README.md')} {os.path.join(packing_package_path, 'README.md')}")
+    package_path = os.path.join(home_path, name)
+    # Check if there is modification after last update. 
+    last_mod_time = find_last_modification(package_path)
     
-    # read the __init__ file and get meta data
-    with open(os.path.join(packing_package_path, name, "__init__.py")) as fp: code = fp.read()
+    # Read the __init__ file and get meta data. 
+    with open(os.path.join(package_path, "__init__.py")) as fp: code = fp.read()
     pivot = "info_manager("
     s = touch(lambda: code.index(pivot))
     if s is None: raise Error("Pack")("Please use info_manager in `__init__.py` to identify the basic information. ")
     info_str = enclosed_object(code, start=s)
     info = info_manager.parse(info_str)
+    last_version = info.get('version', '1.0.0')
+    # Check if there is modification after last update. 
+    if hasattr(info, 'update') and last_mod_time < time.mktime(time.strptime(info.update, update_format)) + 10: return ""
+    # Upload only if the package is modified. 
     info.version_update()
     code = code.replace(info_str, str(info))
-    with open(os.path.join(home_path, name, "__init__.py"), 'w') as fp: fp.write(code)
-    with open(os.path.join(packing_package_path, name, "__init__.py"), 'w') as fp: fp.write(code)
+    code = code.replace(f"__version__ = '{last_version}'", f"__version__ = '{info['version']}'")
+    update_code_list.append((os.path.join(package_path, "__init__.py"), code))
+    import_all(name)
+    missing_require = check_import(name)
+    if missing_require: print(f"Warning: package {name} imported {missing_require} but didnot declare requirement. ")
+
+    # Move package to the temporary working directory. 
+    os.system(f"cp -r {package_path} {os.path.join(packing_package_path, name)}")
+    os.system(f"cp {os.path.join(package_path, 'README.md')} {os.path.join(packing_package_path, 'README.md')}")
+    os.system(f"cp {os.path.join(package_path, 'MANIFEST.in')} {os.path.join(packing_package_path, 'MANIFEST.in')}")
 
     # create the setup.py file
-    setup_file_path = os.path.join(packing_package_path, "setup.py")
+    setup_file_path = os.path.join(packing_package_path, f"setup_{name}.py")
     setup_file_content = "from setuptools import setup, find_packages\n\n"
     options = dict(
         name = name,
@@ -66,7 +100,7 @@ def do_package(name, version=None, pack_type="zipwheel", upload=True):
         author_email = info.get('email', info.get('contact', "")),
 
         packages = "[CODE]find_packages()",
-        include_package_data = "[CODE]True",
+        include_package_data = f"[CODE]{info.get('package_data', False)}",
         platforms = "any",
         install_requires = [x.strip() for x in to_list(info.get('requires', "")) if x.strip()]
     )
@@ -74,27 +108,32 @@ def do_package(name, version=None, pack_type="zipwheel", upload=True):
 
     # pack the package and record the path
     open(setup_file_path, "w").write(setup_file_content)
-    os.system(f"cd {packing_package_path}; python3 setup.py {pack_type_command[pack_type].format(dist_name = name)}")
+    os.system(f"cd {packing_package_path}; python3 setup_{name}.py {pack_type_command[pack_type].format(dist_name = name)}")
         
     return f" {packing_package_path}/dist_{name}/*"
 
-if __name__ == "__main__":
-    if len(sys.argv) > 1: package_names = sys.argv[1:]
-    else: package_names = ['all']
+def pack(*package_names):
+    if len(package_names) == 1 and isinstance(package_names[0], (list, tuple)):
+        package_names = package_names[0]
+    if len(package_names) == 0: package_names = ['all']
 
-    # pack the packages
-    command = "twine upload"
+    # Pack the packages. 
+    if is_valid_command('twine'): command = "twine upload"
+    elif is_valid_command('python3') and is_valid_command('python3 -m twine', "No module named"): command = "python3 -m twine upload"
+    elif is_valid_command('python2') and is_valid_command('python2 -m twine', "No module named"): command = "python2 -m twine upload"
+    elif is_valid_command('python') and is_valid_command('python -m twine', "No module named"): command = "python -m twine upload"
+    else: raise TypeError("Cannot find installed command 'twine'. ")
     if os.path.exists(packing_package_path): os.system(f"rm -r {packing_package_path}")
     os.mkdir(packing_package_path)
     if 'all' in package_names:
         if len(package_names) > 1:
             print("Warning: Indicator 'all' was found as long as ordinary package names. ")
-            r = input("Shall we go on? (Y/N)")
+            r = input("Shall we go on for all available packages? (Y/N)")
             if 'y' not in r.lower(): exit()
         for path in os.listdir(home_path):
             cpath = os.path.join(home_path, path)
             if os.path.isdir(cpath):
-                if "__init__.py" in os.listdir(cpath):
+                if "__init__.py" in os.listdir(cpath) and ".ignore_pack" not in os.listdir(cpath):
                     command += do_package(path)
     else:
         for pn in package_names:
@@ -104,6 +143,26 @@ if __name__ == "__main__":
                 command += do_package(pn, version=v)
             else: print(f"Warning: Package '{pn}' not found. ")
 
-    # upload using twine
-    os.system(command)
+    # # Test command. 
+    # print(command)
+    # return
+    # Upload using twine. 
+    if command.strip().endswith('upload'): print("All mentioned packages are up to date. "); return
+    p = pexpect.spawn(command)
+    p.expect("Enter your username:")
+    p.sendline(key_chain['usr'])
+    p.expect("Enter your password:")
+    p.sendline(key_chain['psw'])
+    p.expect(pexpect.EOF)
+    if b"403 Forbidden" in p.before: print("Upload Failed..."); return
+    if b"400 Bad Request" in p.before: print("Upload Failed (A copy of given version found on the server, please correct the version manually)..."); return
+    else:
+        names = []
+        for f, c in update_code_list:
+            with open(f, 'w') as fp: fp.write(c)
+            names.append(os.path.basename(os.path.dirname(f)))
+        print("Uploaded Successfully:", ', '.join(names));
     os.system(f"rm -r {packing_package_path}")
+
+if __name__ == "__main__":
+    pack(sys.argv[1:])
